+++ Editor/Authoring/AssemblyInfo.cs
@@ -5,5 +5,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.Services.Multiplay.Authoring.IntegrationTests.EditMode")]
 [assembly: InternalsVisibleTo("InternalsVisible.DynamicProxyGenAssembly2")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
-[assembly: InternalsVisibleTo("Unity.DedicatedServer.DeploymentOrchestrator.Editor.MultiplayNode")]
 #endif