+++ Editor/Authoring/MultiplayAuthoringServices.cs
@@ -27,8 +27,14 @@ using IAccessTokens = Unity.Services.Core.Editor.IAccessTokens;
 
 namespace Unity.Services.Multiplay.Authoring.Editor
 {
-    static class MultiplayAuthoringServices
+    /// <summary>
+    /// Provides access to the different services used for the Admin APIs and Deployment capabilities
+    /// </summary>
+    public static class MultiplayAuthoringServices
     {
+        /// <summary>
+        /// The service provider
+        /// </summary>
         public static IScopedServiceProvider Provider { get; }
 
         static MultiplayAuthoringServices()
@@ -57,6 +63,7 @@ namespace Unity.Services.Multiplay.Authoring.Editor
             collection.Register(Default<IBuildConfigApiFactory, MultiplayApiFactory>);
             collection.Register(Default<IAllocationApiFactory, MultiplayApiFactory>);
             collection.Register(Default<IServersApiFactory, MultiplayApiFactory>);
+            collection.Register(Default<ILogsApiFactory, MultiplayApiFactory>);
 
             collection.Register(Default<IAccessTokens, AccessTokens>);
             collection.Register(Default<ICurrentTime, CurrentTime>);
@@ -97,7 +104,10 @@ namespace Unity.Services.Multiplay.Authoring.Editor
             collection.RegisterStartupSingleton(Default<FleetStatusTracker>);
         }
 
-        internal static T GetService<T>(this IServiceProvider provider)
+        /// <summary>
+        /// Get the service specified by the type T
+        /// </summary>
+        public static T GetService<T>(this IServiceProvider provider)
         {
             return (T)provider.GetService(typeof(T));
         }