+++ Runtime/Authoring/Core/Assets/FleetName.cs
@@ -1,8 +1,18 @@
 namespace Unity.Services.Multiplay.Authoring.Core.Assets
 {
-    readonly struct FleetName : IResourceName
+    /// <summary>
+    /// Struct representing a type-safe Fleet Identifier
+    /// </summary>
+    public readonly struct FleetName : IResourceName
     {
+       /// <summary>
+       /// Creates a new instance of FleetName
+       /// </summary>
         public FleetName(string name) { Name = name; }
+
+        /// <summary>
+        /// The name of the Fleet
+        /// </summary>
         public string Name { get; init; }
 
         public override string ToString()