+++ Runtime/Authoring/Core/Builds/CloudBucket.cs
@@ -2,11 +2,21 @@ using System;
 
 namespace Unity.Services.Multiplay.Authoring.Core.Builds
 {
-    struct CloudBucketId
+    /// <summary>
+    /// Represents the identifier of the CCD bucket where direct file upload is stored
+    /// </summary>
+    public struct CloudBucketId
     {
+        /// <summary>
+        /// ID of the cloud bucket
+        /// </summary>
         public Guid Id { private get; init; }
 
+        /// <summary>
+        /// Gets the CloudBucket as a GUID
+        /// </summary>
         public Guid ToGuid() => Id;
+        /// <inheritdoc />
         public override string ToString() => Id.ToString();
     }
 }