+++ Runtime/Authoring/Core/Deployment/DeploymentFacadeFactory.cs
@@ -1,8 +1,8 @@
 using System.Threading.Tasks;
 using Unity.Services.Multiplay.Authoring.Core.Builds;
-using Unity.Services.Multiplay.Authoring.Core.Logging;
 using Unity.Services.Multiplay.Authoring.Core.MultiplayApi;
 using Unity.Services.Multiplay.Authoring.Core.Threading;
+using ILogger = Unity.Services.Multiplay.Authoring.Core.Logging.ILogger;
 
 namespace Unity.Services.Multiplay.Authoring.Core.Deployment
 {