+++ Runtime/Authoring/Core/MultiplayApi/BuildConfigurationId.cs
@@ -1,9 +1,18 @@
 namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
 {
-    struct BuildConfigurationId
+    /// <summary>
+    /// Represents a type-safe Build Configuration ID
+    /// </summary>
+    public struct BuildConfigurationId
     {
+        /// <summary>
+        /// The numerical value of the Build Config TBA
+        /// </summary>
         public long Id { private get; init; }
 
+        /// <summary>
+        /// The numerical value of the Build Config TBA
+        /// </summary>
         public long ToLong() => Id;
     }
 }