+++ Runtime/Authoring/Core/MultiplayApi/BuildId.cs
@@ -1,9 +1,18 @@
 namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
 {
-    struct BuildId
+    /// <summary>
+    /// Represents a type-safe Build ID
+    /// </summary>
+    public struct BuildId
     {
+        /// <summary>
+        /// The numerical value of the Build TBA
+        /// </summary>
         public long Id { private get; init; }
 
+        /// <summary>
+        /// The numerical value of the Build
+        /// </summary>
         public long ToLong() => Id;
     }
 }