+++ Runtime/Authoring/Core/MultiplayApi/IServersApiFactory.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 
 namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
 {
-    interface IServersApiFactory
+    public interface IServersApiFactory
     {
         Task<IServersApi> Build();
     }