+++ Runtime/Authoring/Core/MultiplayApi/ServerInfo.cs
@@ -1,0 +1,13 @@
+namespace Unity.Services.Multiplay.Authoring.Core.MultiplayApi
+{
+    public enum ServerStatus
+    {
+        ALLOCATED = 1,
+        RESERVED = 2,
+        AVAILABLE = 3,
+        ONLINE = 4,
+        READY = 5,
+        HELD = 6
+    }
+    public record ServerInfo(long Id, long MachineID, string MachineName, long BuildConfigurationID, string BuildConfigurationName, string BuildName, System.Guid FleetID, string FleetName, long LocationID, string LocationName, string Ip, int Port, ServerStatus Status, long CpuLimit, long MemoryLimit, bool Deleted, long HoldExpiresAt = default);
+}