+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0-exp.1] - 2023-04-01
+* Added API to interact with Admin API
+* Added API to interact with deployment capabilities
+
+
 ## [1.1.1] - 2023-11-08
 * When creating a build, avoid switching target as this can cause a domain reload, instead emit a warning